Reading and merging, performing a regression, and plotting the data
It is often unclear if two data series, e.g. time series, are statistically related. A first indication
can be gathered from a visual inspection of the two data series plotted against each other.
In this assignment, you will be dealing with data for the growth of real estate prices (Real
Residential Property Prices) in (1) Germany and (2) Austria. The two data series can be
downloaded as CSV files from the website of the Federal Reserve Bank of St. Louis
• https://fred.stlouisfed.org/series/QDER368BIS (Germany)
• https://fred.stlouisfed.org/series/QATR368BIS (Austria)
Problem
1. Please write a Python script for investigating whether there is a statistical relation between the two data series and plot the data and a regression line. Hint: This may involve reading and merging the two data sets, performing a regression, and plotting the data.