Add files via upload
In this assignment, it shall be investigated, how the distribution of wealth changes if the money supply is not constant, namely, if agents invest and get random returns (positive or even negative) between time periods. Random exchanges of money still happen. Agents start with fixed identical amounts of money (1000 monetary units).
The returns should follow the following process
mt+1 = 0.9αtmt + βt 
where m is the agent’s money and αt and βt are uncorrelated, non-autocorrelated, exponentially distributed random variables with αt ∼ 0.92 + Exponential(0.2) and βt ∼ Exponential(0.3).
The sequence of events in each timestep should be:
1. Each agent receives (positive or negative returns on their money following equation 1.
2. Random pairs of agents A and B should be selected for exchanging money (i.e., one gives money to the other), such that each agent participates on average once every ca. 5 timesteps (e.g., if we have 500 agents, we select 50 pairs per timestep). In each pair, agent A should transfer a random (uniformly distributed) share of their current money to agent B.
This should be repeated for at least 1000 timesteps.
Please consider using the code on the following page for your simulation; the return process and various other useful aspects are already implemented.

Please write a script to conduct this simulation. The script should collect and visualize the simulation output by calling the assess_money_distribution method of the Simulation class in the end. Consider using the code below.
Hint: Please note that the constructor method of the Simulation class has not been completed and that the Simulation class will most likely also need a run method.