Add files via upload
The type of interaction network can have a profound effect on the dynamics in social, economic, and other systems. The code below defines a SIR model of an epidemic in a population with an Erd ̋os-R ́enyi interaction network with parameter p = 0.1 (probability of the existence of a link between any two nodes).
In this assignment, the effect of the value of the parameter p on the outcome shall be investigated. Does it have an effect? If so, what effect does it have?
Problem
Write a script for performing the experiment defined in problem 1 (i.e., for investigating the effect of the parameter value p).
Hint: You can start with the script below and modify the script accordingly.